refactor: 회원 탈퇴시 회원의 카드도 연관 삭제 추가

+++ src/main/java/com/sparta/springtodolist/domain/user/service/UserService.java
@@ -1,14 +1,12 @@
 package com.sparta.springtodolist.domain.user.service;
 
-import com.sparta.springtodolist.domain.card.entity.Card;
-import com.sparta.springtodolist.domain.card.repository.CardRepository;
+import com.sparta.springtodolist.domain.card.service.CardService;
 import com.sparta.springtodolist.domain.user.entity.User;
 import com.sparta.springtodolist.domain.user.exception.ExistsUserException;
 import com.sparta.springtodolist.domain.user.repository.UserRepository;
 import com.sparta.springtodolist.domain.user.service.dto.request.UserSignupServiceRequestDto;
 import com.sparta.springtodolist.domain.user.service.dto.response.UserSignupResponseDto;
 import com.sparta.springtodolist.global.exception.ErrorCode;
-import java.util.List;
 import lombok.RequiredArgsConstructor;
 import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.stereotype.Service;
@@ -19,8 +17,8 @@ import org.springframework.transaction.annotation.Transactional;
 @Transactional(readOnly = true)
 public class UserService {
 
+    private final CardService cardService;
     private final UserRepository userRepository;
-    private final CardRepository cardRepository;
     private final PasswordEncoder passwordEncoder;
 
     @Transactional
@@ -38,8 +36,10 @@ public class UserService {
 
     @Transactional
     public void deleteUser(User user) {
-        List<Card> cardList = cardRepository.findByUser(user);
-        cardRepository.deleteAllInBatch(cardList);
+        if (!userRepository.existsById(user.getId())) {
+            throw new ExistsUserException(ErrorCode.EXISTS_USERNAME);
+        }
+        cardService.deleteByUserId(user.getId());
 
         userRepository.delete(user);
     }
